requirements.txt
# --- Core Agent & LLM ---
uagents
openai
langchain
langchain-openai
langchain-community

# --- Data Scraping & Processing ---
requests
beautifulsoup4
markdown

# Selenium for browser automation
selenium
webdriver-manager

# HTML to Markdown conversion
markdownify

# --- Vector Database ---
supabase
psycopg2-binary

# --- Web3 Protocols ---
# For Hedera Consensus Service
hedera-sdk-py

# For interacting with The Graph's GraphQL API
gql[requests]

# For general Ethereum interactions if needed web3

# --- Utilities ---
python-dotenv
uvicorn